requirements.txt
pycryptodome
mutagen